Move utc_boundaries helper around
