+++ lib/plausible/stats/sql/where_builder.ex
@@ -4,7 +4,7 @@ defmodule Plausible.Stats.SQL.WhereBuilder do
   """
 
   import Ecto.Query
-  import Plausible.Stats.Base, only: [utc_boundaries: 2]
+  import Plausible.Stats.Time, only: [utc_boundaries: 2]
   import Plausible.Stats.Filters.Utils, only: [page_regex: 1]
 
   alias Plausible.Stats.Query
