+++ test/plausible/stats/query_test.exs
@@ -21,11 +21,11 @@ defmodule Plausible.Stats.QueryTest do
   } do
     q1 = %{now: %NaiveDateTime{}} = Query.from(site, %{"period" => "realtime"})
     q2 = %{now: %NaiveDateTime{}} = Query.from(site, %{"period" => "30m"})
-    boundaries1 = Plausible.Stats.Base.utc_boundaries(q1, site)
-    boundaries2 = Plausible.Stats.Base.utc_boundaries(q2, site)
+    boundaries1 = Plausible.Stats.Time.utc_boundaries(q1, site)
+    boundaries2 = Plausible.Stats.Time.utc_boundaries(q2, site)
     :timer.sleep(1500)
-    assert ^boundaries1 = Plausible.Stats.Base.utc_boundaries(q1, site)
-    assert ^boundaries2 = Plausible.Stats.Base.utc_boundaries(q2, site)
+    assert ^boundaries1 = Plausible.Stats.Time.utc_boundaries(q1, site)
+    assert ^boundaries2 = Plausible.Stats.Time.utc_boundaries(q2, site)
   end
 
   test "parses day format", %{site: site} do
